backend/requirements.txt
Django==4.2
django-environ==0.11.2
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-cors-headers==4.2.0